+++ src/wrapper.ts
@@ -4,22 +4,18 @@ import type { Monad, Comonad } from './types';
 class Wrapper<T> implements Comonad<T>, Monad<T> {
   // TODO: review this when ECMAScript's private class fields will be
   // widely spread in browsers.
-  private readonly _value: T;
-
-  private constructor(value: T) {
-    this._value = value;
-  }
+  private constructor(private readonly _value: T) {}
 
   /**
    * Wraps value in `Wrapper` monad and allow perform on it operations in chainable way.
    * If value is `Wrapper`, then its copy will be returned.
    */
   static wrap<U>(value: U | Wrapper<U>): Wrapper<U> {
-    return new Wrapper(isWrapper<U>(value) ? value.extract() : value);
+    return new Wrapper<U>(isWrapper<U>(value) ? value.extract() : value);
   }
 
   map<R>(fn: (value: T) => R): Wrapper<R> {
-    return Wrapper.wrap(fn(this._value));
+    return new Wrapper<R>(fn(this._value));
   }
 
   apply<R>(other: Wrapper<(value: T) => R>): Wrapper<R> {